toolsies: Complete Spawn rework
There will be no such thing as unavailable spawn here anymore.
If there is no custom spawn point set, we'll default to getWorlds().get(0)'s location.
 * Fix Component Placeholders
 * Refactor /getspawn command
 * Remove distance, because this was not an useful information
 * /spawn now does not support worlds, it was unnecessary
   Remove the bypass logic along with it
 * Rework PlayerRespawnListener accordingly.
 * Remove Spawn object.
 * Remove /delspawn.
 * Prioritize respawning in bed location - vanilla behaviour.
 * Keep tLocation for future use.

To conclude.
Now, there will be only one default spawn, which will be the "world".
/setspawn command changes the default spawnpoint of the world
and allows the location to be set as preferred if executed twice in the same location.
Preferred spawns work only on death, the /spawn command will use the default location.

 * Added a very bad function to check if String is double. Sorry not sorry.